Exchange cart id input field with a select field
This will prevent trying to insert an device with a none existing cart. At the moment if someone wants to insert a new device with a non existing cart, the query will not be executed due to no matching foreign key. Also values of the cart select field needs to be filled with existing carts of the database in the future.